@@ -6,58 +6,63 @@
   "dependencies": {
     "animate.css": "^3.6.1",
     "babel-plugin-lodash": "^3.3.2",
+    "core-js": "^2.6.0",
     "disqus-react": "^1.0.5",
     "font-awesome": "^4.7.0",
-    "gatsby": "^1.9.149",
-    "gatsby-image": "^1.0.53",
-    "gatsby-link": "^1.6.32",
-    "gatsby-plugin-catch-links": "^1.0.14",
-    "gatsby-plugin-feed": "^1.3.15",
-    "gatsby-plugin-google-analytics": "^1.0.14",
-    "gatsby-plugin-lodash": "^1.0.11",
+    "gatsby": "2.0.40",
+    "gatsby-image": "2.0.19",
+    "gatsby-plugin-catch-links": "2.0.6",
+    "gatsby-plugin-feed": "2.0.9",
+    "gatsby-plugin-google-analytics": "2.0.7",
+    "gatsby-plugin-layout": "^1.0.10",
+    "gatsby-plugin-lodash": "3.0.2",
     "gatsby-plugin-mailchimp": "^2.2.3",
-    "gatsby-plugin-manifest": "^1.0.12",
-    "gatsby-plugin-netlify": "^1.0.21",
-    "gatsby-plugin-netlify-cms": "^2.0.1",
-    "gatsby-plugin-nprogress": "^1.0.9",
-    "gatsby-plugin-offline": "^1.0.12",
-    "gatsby-plugin-page-transitions": "^1.0.7",
-    "gatsby-plugin-react-helmet": "^2.0.3",
-    "gatsby-plugin-sass": "^1.0.15",
-    "gatsby-plugin-sharp": "^1.6.47",
-    "gatsby-plugin-sitemap": "^1.2.9",
-    "gatsby-plugin-twitter": "^1.0.14",
-    "gatsby-remark-autolink-headers": "^1.4.11",
-    "gatsby-remark-copy-linked-files": "^1.5.25",
-    "gatsby-remark-images": "^1.5.36",
-    "gatsby-remark-prismjs": "^1.2.11",
-    "gatsby-remark-responsive-iframe": "^1.4.16",
-    "gatsby-source-filesystem": "^1.5.11",
-    "gatsby-transformer-remark": "^1.7.26",
-    "gatsby-transformer-sharp": "^1.6.26",
+    "gatsby-plugin-manifest": "2.0.7",
+    "gatsby-plugin-netlify": "^2.0.6",
+    "gatsby-plugin-netlify-cms": "3.0.6",
+    "gatsby-plugin-nprogress": "2.0.6",
+    "gatsby-plugin-offline": "2.0.12",
+    "gatsby-plugin-react-helmet": "3.0.1",
+    "gatsby-plugin-sass": "2.0.3",
+    "gatsby-plugin-sharp": "2.0.12",
+    "gatsby-plugin-sitemap": "2.0.2",
+    "gatsby-plugin-twitter": "2.0.7",
+    "gatsby-remark-autolink-headers": "2.0.10",
+    "gatsby-remark-copy-linked-files": "2.0.6",
+    "gatsby-remark-images": "2.0.6",
+    "gatsby-remark-prismjs": "3.0.3",
+    "gatsby-remark-relative-images": "0.2.0",
+    "gatsby-remark-responsive-iframe": "2.0.6",
+    "gatsby-source-filesystem": "2.0.7",
+    "gatsby-transformer-remark": "2.1.11",
+    "gatsby-transformer-sharp": "2.1.8",
     "lodash": "^4.17.10",
     "lodash-webpack-plugin": "^0.11.4",
     "marked": "^0.4.0",
     "netlify-cms": "^1.9.2",
-    "react": "^16.5.0",
-    "react-dom": "^16.5.0",
+    "node-sass": "^4.10.0",
+    "prismjs": "^1.15.0",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
     "react-helmet": "^5.2.0",
-    "react-md": "^1.5.0",
+    "react-md": "^1.8.0",
     "react-particle-effect-button": "^1.0.0",
     "react-share": "^2.1.0",
     "react-twitter-widgets": "^1.7.1",
-    "typed.js": "^2.0.8"
+    "typed.js": "^2.0.8",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "url-join": "^4.0.0"
   },
   "devDependencies": {
     "cli-glob": "^0.1.0",
     "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^3.1.0",
     "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.5.1",
     "gh-pages": "^1.1.0",
-    "prettier": "^1.9.2",
+    "prettier": "^1.14.3",
     "remark-cli": "^5.0.0",
     "remark-preset-lint-recommended": "^3.0.1",
     "stylefmt": "^6.0.0",
