@@ -1,14 +1,14 @@
-module.exports = [
+export default [
   {
     id: 1,
-    link: "https://www.youtube.com/watch?v=MsTCsmqkezQ&amp;t=1",
-    title: "Nothing Funny",
+    link: 'https://www.youtube.com/watch?v=MsTCsmqkezQ&amp;t=1',
+    title: 'Nothing Funny',
     value: `Eric Thomas shows us the real reason he speaks and the power of using your platform to change the lives of those who need hope! Honesty is sometimes the best policy.`
   },
   {
     id: 2,
-    link: "https://www.youtube.com/watch?v=7Oxz060iedY",
-    title: "You Owe You",
+    link: 'https://www.youtube.com/watch?v=7Oxz060iedY',
+    title: 'You Owe You',
     value: `How many times have you made a guarantee to yourself in your life?
             We will quickly call a company after their product breaks before
             the guarantee, but what do you do when you break your own
@@ -16,21 +16,20 @@ module.exports = [
   },
   {
     id: 3,
-    link:
-      "https://www.joshuakennon.com/give-yourself-permission-to-be-disliked/",
-    title: "Give Yourself Permission to be Disliked",
+    link: 'https://www.joshuakennon.com/give-yourself-permission-to-be-disliked/',
+    title: 'Give Yourself Permission to be Disliked',
     value: `This article has changed my perspective on the world - what an incredible view and  honest article.`
   },
   {
     id: 4,
-    link: "https://www.youtube.com/watch?v=5U0TE4oqj24",
-    title: "Charlie Munger USC Commencement Speech",
+    link: 'https://www.youtube.com/watch?v=5U0TE4oqj24',
+    title: 'Charlie Munger USC Commencement Speech',
     value: `A video of one of the greatest thinkers giving keys to a good life.`
   },
   {
     id: 5,
-    link: "https://www.youtube.com/watch?v=pxBQLFLei70",
-    title: "University of Texas at Austin 2014 Commencement Address - Admiral William H. McRaven",
-    "value": `Make your bed. Don't ever, ever ring the bell!`
+    link: 'https://www.youtube.com/watch?v=pxBQLFLei70',
+    title: 'University of Texas at Austin 2014 Commencement Address - Admiral William H. McRaven',
+    value: `Make your bed. Don't ever, ever ring the bell!`
   }
 ];
