@@ -1,7 +1,7 @@
-module.exports = [
+export default [
   {
     id: 1,
-    title: "Relationships and You",
+    title: 'Relationships and You',
     value: `Many positives are commonly mentioned about
     relationships - support, love, friendship, and more - while one I
     believe is left out. That one thing? The amount you grow and learn
@@ -10,7 +10,7 @@ module.exports = [
   },
   {
     id: 2,
-    title: "Doing Something Lessons",
+    title: 'Doing Something Lessons',
     value: `Sometimes just doing something before you
     actually have the full plan ready results in better focus to grow
     the plan, or fail fast on a project you don't
@@ -19,7 +19,7 @@ module.exports = [
   },
   {
     id: 3,
-    title: "Habits are key",
+    title: 'Habits are key',
     value: `The person you are, the success you bring or
     happiness you have doesn't come in spikes. Instead, they are like
     compound interest, gradually growing into the high levels you
@@ -29,37 +29,37 @@ module.exports = [
   },
   {
     id: 4,
-    title: "Actions Always Speak Like Pictures Describe",
+    title: 'Actions Always Speak Like Pictures Describe',
     value: `People in your life can tell you how much they respect you, care about you, appreciate you, etc. These declarations are great and an important part of relationships, yet it is always the actions that either dispel or backup these statements that truly matter.`
   },
   {
     id: 5,
-    title: "Spending Time > Spending Money",
+    title: 'Spending Time > Spending Money',
     value:
-      "You can always do more to increase the amount of money you have in bank accounts, but there is an account that decreases every second. Your time balance. There is nothing more important than giving the people you love in your life some of those ticking seconds."
+      'You can always do more to increase the amount of money you have in bank accounts, but there is an account that decreases every second. Your time balance. There is nothing more important than giving the people you love in your life some of those ticking seconds.'
   },
   {
     id: 6,
-    title: "Measurements in Your Life",
+    title: 'Measurements in Your Life',
     value:
-      "We each use different measurements for other people in our life, friendships or spouses as we want to have balance within those relationships. One key point to remember? You can’t control others and if they don’t have the same measurements in place – you may end up disappointed. Communicate those expectations if necessary, or accept them for who they are. The boundary should be of their character, not the measurements of their contact or other expectations you set on them."
+      'We each use different measurements for other people in our life, friendships or spouses as we want to have balance within those relationships. One key point to remember? You can’t control others and if they don’t have the same measurements in place – you may end up disappointed. Communicate those expectations if necessary, or accept them for who they are. The boundary should be of their character, not the measurements of their contact or other expectations you set on them.'
   },
   {
     id: 7,
-    title: "Social Media Doesn’t Paint the Truth",
+    title: 'Social Media Doesn’t Paint the Truth',
     value:
       "Taking a peek at someone’s social media accounts as an indicator of their life doesn’t work. People post in different ways, only show specific information, and there isn't a true connection in a page visit. Real friendships are those who catch up through face to face or phone calls that produce intimate conversations to build trust. There is no other way."
   },
   {
     id: 8,
-    title: "Imposter Syndrome is Real",
+    title: 'Imposter Syndrome is Real',
     value:
-      "Expertise is a fluid concept that doesn’t have defined expectations. In reality, most of the time, there is someone who has more expertise than you on a subject. Let them be and live your life in peace. Instead, do your best to get a little smarter, and a little wiser each day."
+      'Expertise is a fluid concept that doesn’t have defined expectations. In reality, most of the time, there is someone who has more expertise than you on a subject. Let them be and live your life in peace. Instead, do your best to get a little smarter, and a little wiser each day.'
   },
   {
     id: 9,
-    title: "A Key to a Good Life is Avoiding Big Mistakes",
+    title: 'A Key to a Good Life is Avoiding Big Mistakes',
     value:
-      "Mistakes are going to happen in your life, there is no mistaking that. Decide where you want to be at the end of your life and start using inversion to see which actions take you there. For each action, ask yourself – what would be detrimintal to reaching these goals….avoid them like the plague, or bitcoin."
+      'Mistakes are going to happen in your life, there is no mistaking that. Decide where you want to be at the end of your life and start using inversion to see which actions take you there. For each action, ask yourself – what would be detrimintal to reaching these goals….avoid them like the plague, or bitcoin.'
   }
 ];
