@@ -1,148 +1,143 @@
-module.exports = {
+export default {
   // need to add rocket, figma, computer, trello, and other productivity flows
   // books/podcasts
   dev: [
     {
-      category: "Web Development",
+      category: 'Web Development',
       subcategories: [
         {
-          name: "Javascript Frameworks/Libraries",
+          name: 'Javascript Frameworks/Libraries',
           resources: [
             {
-              name: "React",
-              link: "https://reactjs.org/",
+              name: 'React',
+              link: 'https://reactjs.org/',
               description: "Facebook's library to build user interfaces."
             },
             {
-              name: "Lodash",
-              link: "https://lodash.com/docs/",
-              description: "Performant, re-usable utility library."
+              name: 'Lodash',
+              link: 'https://lodash.com/docs/',
+              description: 'Performant, re-usable utility library.'
             },
             {
-              name: "MomentJS",
-              link: "https://momentjs.com/",
-              description: "Makes dates, time, and durations a breeze."
+              name: 'MomentJS',
+              link: 'https://momentjs.com/',
+              description: 'Makes dates, time, and durations a breeze.'
             },
             {
-              name: "Redux",
-              link: "https://redux.js.org/",
-              description: "Manage your application state in one place."
+              name: 'Redux',
+              link: 'https://redux.js.org/',
+              description: 'Manage your application state in one place.'
             },
             {
-              name: "Gatsby",
-              link: "https://gatsby.org",
-              description: "React / GraphQL static site generator."
+              name: 'Gatsby',
+              link: 'https://gatsby.org',
+              description: 'React / GraphQL static site generator.'
             },
             {
-              name: "React MD",
-              link: "https://react-md.mlaursen.com/",
-              description: "React Material Design Components"
+              name: 'React MD',
+              link: 'https://react-md.mlaursen.com/',
+              description: 'React Material Design Components'
             }
           ]
         },
         {
-          name: "CSS Libraries",
+          name: 'CSS Libraries',
           resources: [
             {
-              name: "Bootstrap",
-              link: "https://getbootstrap.com/",
-              description: "Toolkit to make your front-ends responsive."
+              name: 'Bootstrap',
+              link: 'https://getbootstrap.com/',
+              description: 'Toolkit to make your front-ends responsive.'
             },
             {
-              name: "Bulma",
-              link: "https://bulma.io/",
-              description:
-                "CSS framework based on flexbox to make responsive layouts easy"
+              name: 'Bulma',
+              link: 'https://bulma.io/',
+              description: 'CSS framework based on flexbox to make responsive layouts easy'
             },
             {
-              name: "Sass",
-              link: "https://sass-lang.com/",
-              description: "Extends CSS with synatic sugar that makes sense."
+              name: 'Sass',
+              link: 'https://sass-lang.com/',
+              description: 'Extends CSS with synatic sugar that makes sense.'
             },
             {
-              name: "Animate",
-              link: "https://daneden.github.io/animate.css/",
-              description: "Add easy animations via class names."
+              name: 'Animate',
+              link: 'https://daneden.github.io/animate.css/',
+              description: 'Add easy animations via class names.'
             },
             {
-              name: "Lottie",
-              link: "https://airbnb.design/lottie/",
+              name: 'Lottie',
+              link: 'https://airbnb.design/lottie/',
               description: "AirBnb's library to make animations easy."
             }
           ]
         },
         {
-          name: "Bundlers / Package Managers",
+          name: 'Bundlers / Package Managers',
           resources: [
             {
-              name: "Yarn",
-              link: "https://yarnpkg.com/lang/en/",
-              description: "Ultra fast, secure node package manager."
+              name: 'Yarn',
+              link: 'https://yarnpkg.com/lang/en/',
+              description: 'Ultra fast, secure node package manager.'
             },
             {
-              name: "Webpack",
-              link: "https://webpack.js.org/",
-              description: "Static module bundler for modern JS."
+              name: 'Webpack',
+              link: 'https://webpack.js.org/',
+              description: 'Static module bundler for modern JS.'
             },
             {
-              name: "Parcel",
-              link: "https://parceljs.org/",
-              description:
-                "New competitor for asset bundling - boasts zero config with great bench marks!"
+              name: 'Parcel',
+              link: 'https://parceljs.org/',
+              description: 'New competitor for asset bundling - boasts zero config with great bench marks!'
             }
           ]
         }
       ]
     },
     {
-      category: "Productivity Tools",
+      category: 'Productivity Tools',
       subcategories: [
         {
-          name: "Editor",
+          name: 'Editor',
           resources: [
             {
               name: "Visual Studio Code - Insider's Edition",
-              link: "https://code.visualstudio.com/insiders/",
+              link: 'https://code.visualstudio.com/insiders/',
               description: "Microsoft's Electron based text editor."
             },
             {
-              name: "Visual Studio Code - Settings Sync",
-              link:
-                "https://marketplace.visualstudio.com/items?itemName=Shan.code-settings-sync",
+              name: 'Visual Studio Code - Settings Sync',
+              link: 'https://marketplace.visualstudio.com/items?itemName=Shan.code-settings-sync',
               description:
-                "Easily download/upload to share settings by following the Settings Sync walk through. If interested, use my public gist (baad940d8ae370312e4a748cc3cdc8d6) to stay updated with my recommended theme, settings, extensions! Utilizing Visual Studio Code, you can increase your productivity as a developer!"
+                'Easily download/upload to share settings by following the Settings Sync walk through. If interested, use my public gist (baad940d8ae370312e4a748cc3cdc8d6) to stay updated with my recommended theme, settings, extensions! Utilizing Visual Studio Code, you can increase your productivity as a developer!'
             }
           ]
         },
         {
-          name: "Terminal",
+          name: 'Terminal',
           resources: [
             {
-              name: "ITerm2",
-              link: "https://www.iterm2.com/",
-              description:
-                "Increase your productivity in the terminal with the little things."
+              name: 'ITerm2',
+              link: 'https://www.iterm2.com/',
+              description: 'Increase your productivity in the terminal with the little things.'
             },
             {
-              name: "Oh My Zsh",
-              link: "https://ohmyz.sh/",
+              name: 'Oh My Zsh',
+              link: 'https://ohmyz.sh/',
               description:
-                "Framework for manaing your zsh configuration, letting you easily install plugins and themes."
+                'Framework for manaing your zsh configuration, letting you easily install plugins and themes.'
             }
           ]
         },
         {
-          name: "Mac",
+          name: 'Mac',
           resources: [
             {
-              name: "Alfred 3",
-              link: "https://www.alfredapp.com/",
-              description:
-                "Boost your productivity with hotkeys, text expansion, workflows, and search."
+              name: 'Alfred 3',
+              link: 'https://www.alfredapp.com/',
+              description: 'Boost your productivity with hotkeys, text expansion, workflows, and search.'
             },
             {
-              name: "BetterTouchTool",
-              link: "https://folivora.ai/",
+              name: 'BetterTouchTool',
+              link: 'https://folivora.ai/',
               description:
                 "Customize your mac's gestures, hotkeys, and other. Use this with Karabiner-Elements to map your caps lock to unique key combinations to create your own productivity enhancers!."
             }
@@ -153,277 +148,269 @@ module.exports = {
   ],
   life: [
     {
-      category: "Apps",
+      category: 'Apps',
       subcategories: [
         {
-          name: "All",
+          name: 'All',
           resources: [
             {
-              name: "Fabulous App",
-              link: "https://www.thefabulous.co/",
-              description: "Routine app"
+              name: 'Fabulous App',
+              link: 'https://www.thefabulous.co/',
+              description: 'Routine app'
             },
             {
-              name: "Headspace",
-              link: "https://www.headspace.com/",
-              description: "Meditation app"
+              name: 'Headspace',
+              link: 'https://www.headspace.com/',
+              description: 'Meditation app'
             },
             {
-              name: "JEFIT",
-              link: "https://www.jefit.com/",
-              description: "Workout Tracker"
+              name: 'JEFIT',
+              link: 'https://www.jefit.com/',
+              description: 'Workout Tracker'
             },
             {
-              name: "Mint",
-              link: "https://www.onenote.com/",
-              description: "Budgeting and income tracking"
+              name: 'Mint',
+              link: 'https://www.onenote.com/',
+              description: 'Budgeting and income tracking'
             },
             {
-              name: "OneNote",
-              link: "https://www.onenote.com/",
-              description: "Note taking with cloud"
+              name: 'OneNote',
+              link: 'https://www.onenote.com/',
+              description: 'Note taking with cloud'
             },
             {
-              name: "Podbean",
+              name: 'Podbean',
               link:
-                "http://www.podbean.com/start-podcast?sourceid=goog_66&amp;gclid=Cj0KEQjw6LXIBRCUqIjXmdKBxZUBEiQA_f50Plfbdw8ODcy0AAy7tbnT89Br32zmF8wSpqjYXeSGpoAaAgAs8P8HAQ",
-              description: "Podcasts app for Android"
+                'http://www.podbean.com/start-podcast?sourceid=goog_66&amp;gclid=Cj0KEQjw6LXIBRCUqIjXmdKBxZUBEiQA_f50Plfbdw8ODcy0AAy7tbnT89Br32zmF8wSpqjYXeSGpoAaAgAs8P8HAQ',
+              description: 'Podcasts app for Android'
             },
             {
-              name: "Trello",
-              link: "https://trello.com/",
-              description: "Task manager with boards and cards."
+              name: 'Trello',
+              link: 'https://trello.com/',
+              description: 'Task manager with boards and cards.'
             }
           ]
         }
       ]
     },
     {
-      category: "Books",
+      category: 'Books',
       subcategories: [
         {
-          name: "(Auto)Biography",
+          name: '(Auto)Biography',
           resources: [
             {
               name: "Ben Franklin's Autobiography",
               link:
-                "https://www.amazon.com/Benjamin-Franklin-American-Walter-Isaacson/dp/074325807X/ref=pd_lpo_sbs_14_t_1?_encoding=UTF8&amp;psc=1&amp;refRID=CWGHCC0H0CFNJB916PDE"
+                'https://www.amazon.com/Benjamin-Franklin-American-Walter-Isaacson/dp/074325807X/ref=pd_lpo_sbs_14_t_1?_encoding=UTF8&amp;psc=1&amp;refRID=CWGHCC0H0CFNJB916PDE'
             },
             {
               name: "Poor Charlie's Almanack",
-              link:
-                "https://www.amazon.com/Poor-Charlies-Almanack-Charles-Expanded/dp/1578645018"
+              link: 'https://www.amazon.com/Poor-Charlies-Almanack-Charles-Expanded/dp/1578645018'
             }
           ]
         },
         {
-          name: "Fantasy",
+          name: 'Fantasy',
           resources: [
             {
-              name: "Bartimeus Trilogy",
-              link:
-                "https://www.amazon.com/Bartimaeus-3-book-boxed-set-Trilogy/dp/1423136829"
+              name: 'Bartimeus Trilogy',
+              link: 'https://www.amazon.com/Bartimaeus-3-book-boxed-set-Trilogy/dp/1423136829'
             },
             {
-              name: "Harry Potter",
+              name: 'Harry Potter',
               link:
-                "https://www.amazon.com/Harry-Potter-Paperback-Box-Books/dp/0545162076/ref=sr_1_1?ie=UTF8&amp;qid=1502646670&amp;sr=8-1&amp;keywords=harry+potter+books"
+                'https://www.amazon.com/Harry-Potter-Paperback-Box-Books/dp/0545162076/ref=sr_1_1?ie=UTF8&amp;qid=1502646670&amp;sr=8-1&amp;keywords=harry+potter+books'
             },
             {
-              name: "Inheritance Cylcle",
+              name: 'Inheritance Cylcle',
               link:
-                "https://www.amazon.com/Inheritance-Cycle-4-Book/dp/B00ZJ6W7SI/ref=sr_1_1_ha?s=digital-text&amp;ie=UTF8&amp;qid=1494126056&amp;sr=1-1&amp;keywords=inheritance+cycle"
+                'https://www.amazon.com/Inheritance-Cycle-4-Book/dp/B00ZJ6W7SI/ref=sr_1_1_ha?s=digital-text&amp;ie=UTF8&amp;qid=1494126056&amp;sr=1-1&amp;keywords=inheritance+cycle'
             },
             {
               name: "The Ranger's Apprentice",
-              link:
-                "https://www.amazon.com/Rangers-Apprentice-Complete-Book-Set/dp/B00PHO4HZ8"
+              link: 'https://www.amazon.com/Rangers-Apprentice-Complete-Book-Set/dp/B00PHO4HZ8'
             }
           ]
         },
         {
-          name: "Personal Development",
+          name: 'Personal Development',
           resources: [
             {
-              name: "7 Habits of Highly Effective People",
+              name: '7 Habits of Highly Effective People',
               link:
-                "https://www.amazon.com/Habits-Highly-Effective-People-Powerful/dp/1451639619/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126313&amp;sr=1-1&amp;keywords=7+habits+of+highly+effective+people"
+                'https://www.amazon.com/Habits-Highly-Effective-People-Powerful/dp/1451639619/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126313&amp;sr=1-1&amp;keywords=7+habits+of+highly+effective+people'
             },
             {
-              name: "Getting Things Done",
+              name: 'Getting Things Done',
               link:
-                "https://www.amazon.com/How-Win-Friends-Influence-People/dp/0671027034/ref=pd_bxgy_14_img_2?_encoding=UTF8&amp;pd_rd_i=0671027034&amp;pd_rd_r=3Q782TYE90YTFVGNG6VN&amp;pd_rd_w=KPGtN&amp;pd_rd_wg=lJ7r7&amp;psc=1&amp;refRID=3Q782TYE90YTFVGNG6VN"
+                'https://www.amazon.com/How-Win-Friends-Influence-People/dp/0671027034/ref=pd_bxgy_14_img_2?_encoding=UTF8&amp;pd_rd_i=0671027034&amp;pd_rd_r=3Q782TYE90YTFVGNG6VN&amp;pd_rd_w=KPGtN&amp;pd_rd_wg=lJ7r7&amp;psc=1&amp;refRID=3Q782TYE90YTFVGNG6VN'
             },
             {
-              name: "How to Win Friends and Influence People",
+              name: 'How to Win Friends and Influence People',
               link:
-                "https://www.amazon.com/How-Win-Friends-Influence-People/dp/0671027034/ref=pd_bxgy_14_img_2?_encoding=UTF8&amp;pd_rd_i=0671027034&amp;pd_rd_r=3Q782TYE90YTFVGNG6VN&amp;pd_rd_w=KPGtN&amp;pd_rd_wg=lJ7r7&amp;psc=1&amp;refRID=3Q782TYE90YTFVGNG6VN"
+                'https://www.amazon.com/How-Win-Friends-Influence-People/dp/0671027034/ref=pd_bxgy_14_img_2?_encoding=UTF8&amp;pd_rd_i=0671027034&amp;pd_rd_r=3Q782TYE90YTFVGNG6VN&amp;pd_rd_w=KPGtN&amp;pd_rd_wg=lJ7r7&amp;psc=1&amp;refRID=3Q782TYE90YTFVGNG6VN'
             },
             {
-              name: " IChing Wisdom",
-              link:
-                "https://www.amazon.com/Ching-Wisdom-One-Guidance-Answers/dp/0943015421"
+              name: ' IChing Wisdom',
+              link: 'https://www.amazon.com/Ching-Wisdom-One-Guidance-Answers/dp/0943015421'
             },
             {
-              name: "Leading with Heart",
-              link:
-                "https://www.amazon.com/Leading-Heart-Successful-Strategies-Basketball/dp/0446676780"
+              name: 'Leading with Heart',
+              link: 'https://www.amazon.com/Leading-Heart-Successful-Strategies-Basketball/dp/0446676780'
             },
             {
-              name: "Subtle Art of Not Giving a F***",
-              link:
-                "https://www.amazon.com/Subtle-Art-Not-Giving-Counterintuitive/dp/0062457713"
+              name: 'Subtle Art of Not Giving a F***',
+              link: 'https://www.amazon.com/Subtle-Art-Not-Giving-Counterintuitive/dp/0062457713'
             },
             {
-              name: "The Magic of Thinking Big",
+              name: 'The Magic of Thinking Big',
               link:
-                "https://www.amazon.com/Magic-Thinking-Big-David-Schwartz/dp/0671646788/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126421&amp;sr=1-1&amp;keywords=magic+of+thinking+big"
+                'https://www.amazon.com/Magic-Thinking-Big-David-Schwartz/dp/0671646788/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126421&amp;sr=1-1&amp;keywords=magic+of+thinking+big'
             },
             {
-              name: "Transform",
+              name: 'Transform',
               link:
-                "https://www.amazon.com/TransForm-Dramatically-Improve-Business-Relationships-ebook/dp/B00MTNC6L2/ref=sr_1_fkmr0_1?s=books&amp;ie=UTF8&amp;qid=1494126388&amp;sr=1-1-fkmr0&amp;keywords=transform+john+haden"
+                'https://www.amazon.com/TransForm-Dramatically-Improve-Business-Relationships-ebook/dp/B00MTNC6L2/ref=sr_1_fkmr0_1?s=books&amp;ie=UTF8&amp;qid=1494126388&amp;sr=1-1-fkmr0&amp;keywords=transform+john+haden'
             }
           ]
         },
         {
-          name: "Personal Finance",
+          name: 'Personal Finance',
           resources: [
             {
-              name: "The Complete Investor",
+              name: 'The Complete Investor',
               link:
-                "https://www.amazon.com/Charlie-Munger-Complete-Investor-Publishing/dp/023117098X/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126694&amp;sr=1-1&amp;keywords=the+complete+investor"
+                'https://www.amazon.com/Charlie-Munger-Complete-Investor-Publishing/dp/023117098X/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126694&amp;sr=1-1&amp;keywords=the+complete+investor'
             },
             {
-              name: "The Intelligent Investor",
+              name: 'The Intelligent Investor',
               link:
-                "https://www.amazon.com/Intelligent-Investor-Definitive-Investing-Essentials/dp/0060555661/ref=pd_sim_14_6?_encoding=UTF8&amp;pd_rd_i=0060555661&amp;pd_rd_r=AC8CAKNPDQMPK7BT1GXR&amp;pd_rd_w=nDz7x&amp;pd_rd_wg=Dc1SN&amp;psc=1&amp;refRID=AC8CAKNPDQMPK7BT1GXR"
+                'https://www.amazon.com/Intelligent-Investor-Definitive-Investing-Essentials/dp/0060555661/ref=pd_sim_14_6?_encoding=UTF8&amp;pd_rd_i=0060555661&amp;pd_rd_r=AC8CAKNPDQMPK7BT1GXR&amp;pd_rd_w=nDz7x&amp;pd_rd_wg=Dc1SN&amp;psc=1&amp;refRID=AC8CAKNPDQMPK7BT1GXR'
             },
             {
-              name: "The Millionaire Next Door",
-              link:
-                "https://www.dukeofdollars.com/building-wealth/millionaire-next-door-review/"
+              name: 'The Millionaire Next Door',
+              link: 'https://www.dukeofdollars.com/building-wealth/millionaire-next-door-review/'
             },
             {
-              name: "Unshakeable",
-              link:
-                "https://dukeofdollars.com/building-wealth/unshakeable-review-p1/"
+              name: 'Unshakeable',
+              link: 'https://dukeofdollars.com/building-wealth/unshakeable-review-p1/'
             },
             {
-              name: "Your Life or Your Money",
+              name: 'Your Life or Your Money',
               link:
-                "https://www.amazon.com/Your-Money-Life-Transforming-Relationship/dp/0143115766/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126565&amp;sr=1-1&amp;keywords=your+life+or+your+money"
+                'https://www.amazon.com/Your-Money-Life-Transforming-Relationship/dp/0143115766/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126565&amp;sr=1-1&amp;keywords=your+life+or+your+money'
             }
           ]
         },
         {
-          name: "Programming",
+          name: 'Programming',
           resources: [
             {
-              name: "Code Complete",
+              name: 'Code Complete',
               link:
-                "https://www.amazon.com/Code-Complete-Practical-Handbook-Construction/dp/0735619670/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126829&amp;sr=1-1&amp;keywords=clean+code+2"
+                'https://www.amazon.com/Code-Complete-Practical-Handbook-Construction/dp/0735619670/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126829&amp;sr=1-1&amp;keywords=clean+code+2'
             },
             {
-              name: "Pragmatic Programmer",
+              name: 'Pragmatic Programmer',
               link:
-                "https://www.amazon.com/Pragmatic-Programmer-Journeyman-Master/dp/020161622X/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126792&amp;sr=1-1&amp;keywords=pragmatic+programmer"
+                'https://www.amazon.com/Pragmatic-Programmer-Journeyman-Master/dp/020161622X/ref=sr_1_1?s=books&amp;ie=UTF8&amp;qid=1494126792&amp;sr=1-1&amp;keywords=pragmatic+programmer'
             }
           ]
         }
       ]
     },
     {
-      category: "Podcasts",
+      category: 'Podcasts',
       subcategories: [
         {
-          name: "Personal Development",
+          name: 'Personal Development',
           resources: [
             {
-              name: "Eric Thomas Podcast",
-              link: "https://etinspires.com/blog"
+              name: 'Eric Thomas Podcast',
+              link: 'https://etinspires.com/blog'
             },
             {
-              name: "Good Life Project",
-              link: "http://www.goodlifeproject.com/radio/"
+              name: 'Good Life Project',
+              link: 'http://www.goodlifeproject.com/radio/'
             },
             {
-              name: "Tony Robbins Podcast",
-              link: "https://www.tonyrobbins.com/podcast/"
+              name: 'Tony Robbins Podcast',
+              link: 'https://www.tonyrobbins.com/podcast/'
             }
           ]
         },
         {
-          name: "Technology",
+          name: 'Technology',
           resources: [
             {
-              name: "Coding Blocks",
-              link: "http://kaleb.mckelvey@ge.com"
+              name: 'Coding Blocks',
+              link: 'http://kaleb.mckelvey@ge.com'
             }
           ]
         }
       ]
     },
     {
-      category: "Sites/Blogs",
+      category: 'Sites/Blogs',
       subcategories: [
         {
-          name: "Personal Development",
+          name: 'Personal Development',
           resources: [
             {
-              name: "Project Life Mastery",
-              link: "http://projectlifemastery.com/"
+              name: 'Project Life Mastery',
+              link: 'http://projectlifemastery.com/'
             }
           ]
         },
         {
-          name: "Personal Finance",
+          name: 'Personal Finance',
           resources: [
             {
-              name: "Duke of Dollars",
-              link: "https://dukeofdollars.com"
+              name: 'Duke of Dollars',
+              link: 'https://dukeofdollars.com'
             },
             {
-              name: "Mr. Money Mustache",
-              link: "https://www.mrmoneymustache.com/"
+              name: 'Mr. Money Mustache',
+              link: 'https://www.mrmoneymustache.com/'
             },
             {
-              name: "Reddit Personal Finance",
-              link: "https://www.reddit.com/r/personalfinance/"
+              name: 'Reddit Personal Finance',
+              link: 'https://www.reddit.com/r/personalfinance/'
             },
             {
-              name: "The Green Swan",
-              link: "http://www.thegreenswan.org"
+              name: 'The Green Swan',
+              link: 'http://www.thegreenswan.org'
             }
           ]
         },
         {
-          name: "Technology",
+          name: 'Technology',
           resources: [
             {
-              name: "Engadget",
-              link: "https://engadget.com"
+              name: 'Engadget',
+              link: 'https://engadget.com'
             },
             {
-              name: "Wired",
-              link: "https://wired.com"
+              name: 'Wired',
+              link: 'https://wired.com'
             }
           ]
         },
         {
-          name: "Other",
+          name: 'Other',
           resources: [
             {
-              name: "Bad A** of the Week",
-              link: "http://badassoftheweek.com"
+              name: 'Bad A** of the Week',
+              link: 'http://badassoftheweek.com'
             },
             {
-              name: "Joshua Kennon",
-              link: "https://joshuakennon.com"
+              name: 'Joshua Kennon',
+              link: 'https://joshuakennon.com'
             },
             {
-              name: "Lifehacker",
-              link: "https://lifehacker.com"
+              name: 'Lifehacker',
+              link: 'https://lifehacker.com'
             }
           ]
         }
