@@ -1,17 +1,17 @@
-import ServiceOptions from "./ServiceOptionsData";
+import ServiceOptions from './ServiceOptionsData';
 
 const { general, webApp, wordpress } = ServiceOptions;
 
 // lowest package total price option should match min slider option, because this
 // is how I am determining which package to show based on slider option picked
-module.exports = [
+export default [
   {
-    title: "WordPress Site/Blog",
-    subtitle: "User friendly, professional blog or website!",
+    title: 'WordPress Site/Blog',
+    subtitle: 'User friendly, professional blog or website!',
     description: `<p>The WordPress content management system makes it easy for users to update content, write new blog posts, and manage SEO. Depending on budget, I offer a few different packages that will<strong> take your career or business to the next level!</strong></p>`,
     sliderOptions: {
       showSlider: true,
-      label: "Slide to adjust your budget",
+      label: 'Slide to adjust your budget',
       min: 500,
       max: 800,
       step: 100,
@@ -19,7 +19,7 @@ module.exports = [
     },
     packageOptions: [
       {
-        name: "starter",
+        name: 'starter',
         totalPrice: 500,
         options: [
           general.domain.all,
@@ -31,7 +31,7 @@ module.exports = [
         ]
       },
       {
-        id: "basic",
+        id: 'basic',
         totalPrice: 600,
         options: [
           general.domain.all,
@@ -43,7 +43,7 @@ module.exports = [
         ]
       },
       {
-        id: "basicPro",
+        id: 'basicPro',
         totalPrice: 700,
         options: [
           general.domain.all,
@@ -55,7 +55,7 @@ module.exports = [
         ]
       },
       {
-        id: "custom",
+        id: 'custom',
         totalPrice: 800,
         options: [
           general.domain.all,
@@ -67,7 +67,7 @@ module.exports = [
         ]
       },
       {
-        id: "custom",
+        id: 'custom',
         totalPrice: 600,
         options: [
           general.domain.all,
@@ -80,12 +80,12 @@ module.exports = [
     ]
   },
   {
-    title: "Custom Web Application",
-    subtitle: "Customized web applications for your business",
+    title: 'Custom Web Application',
+    subtitle: 'Customized web applications for your business',
     description: `<p>Have a new business idea that requires custom development?  I follow best practices to develop clean solutions that meet your needs. Testing can be included if you prefer (I do), but does mean more development time.`,
     sliderOptions: {
       showSlider: true,
-      label: "Slide to adjust your budget",
+      label: 'Slide to adjust your budget',
       min: 1500,
       max: 4500,
       step: 1500,
@@ -93,7 +93,7 @@ module.exports = [
     },
     packageOptions: [
       {
-        name: "basic",
+        name: 'basic',
         totalPrice: 1500,
         options: [
           general.domain.all,
@@ -104,7 +104,7 @@ module.exports = [
         ]
       },
       {
-        id: "intermediate",
+        id: 'intermediate',
         totalPrice: 3000,
         options: [
           general.domain.all,
@@ -115,7 +115,7 @@ module.exports = [
         ]
       },
       {
-        id: "complex",
+        id: 'complex',
         totalPrice: 4500,
         options: [
           general.domain.all,
@@ -128,8 +128,8 @@ module.exports = [
     ]
   },
   {
-    title: "Web Development PT Contracting",
-    subtitle: "Developing features to impact your users!",
+    title: 'Web Development PT Contracting',
+    subtitle: 'Developing features to impact your users!',
     description: `<p>Need extra help on your business's website or web application? I implement features to spec by following team standards, workflow, and branching strategies.</p>
 
     <p>I have experience in React, Polymer, Node, Springboot, Sass, and other modern web technologies.`,
@@ -139,8 +139,8 @@ module.exports = [
     packageOptions: []
   },
   {
-    title: "AX Dynamics PT Contracting",
-    subtitle: "Helping you hit project deadlines!",
+    title: 'AX Dynamics PT Contracting',
+    subtitle: 'Helping you hit project deadlines!',
     description: `<p>With over 3 years of experience working with AX 2012 R2, I am here to help with tasks that help improve user workflows through customizations, or help with data migrations and upgrades.</p><p>Have a task? I am here to help!</p>`,
     sliderOptions: {
       showSlider: false
