@@ -1,14 +1,21 @@
-const path = require("path");
-const _ = require("lodash");
-const webpackLodashPlugin = require("lodash-webpack-plugin");
+const path = require('path');
+const _ = require('lodash');
+const moment = require('moment');
+const siteConfig = require('./data/SiteConfig');
 
 const postNodes = [];
 
 function addSiblingNodes(createNodeField) {
-  postNodes.sort(
-    ({ frontmatter: { date: date1 } }, { frontmatter: { date: date2 } }) =>
-      new Date(date1) - new Date(date2)
-  );
+  postNodes.sort(({ frontmatter: { date: date1 } }, { frontmatter: { date: date2 } }) => {
+    const dateA = moment(date1, siteConfig.dateFromFormat);
+    const dateB = moment(date2, siteConfig.dateFromFormat);
+
+    if (dateA.isBefore(dateB)) return 1;
+
+    if (dateB.isBefore(dateA)) return -1;
+
+    return 0;
+  });
   for (let i = 0; i < postNodes.length; i += 1) {
     const nextID = i + 1 < postNodes.length ? i + 1 : 0;
     const prevID = i - 1 > 0 ? i - 1 : postNodes.length - 1;
@@ -17,71 +24,80 @@ function addSiblingNodes(createNodeField) {
     const prevNode = postNodes[prevID];
     createNodeField({
       node: currNode,
-      name: "nextTitle",
+      name: 'nextTitle',
       value: nextNode.frontmatter.title
     });
     createNodeField({
       node: currNode,
-      name: "nextSlug",
+      name: 'nextSlug',
       value: nextNode.fields.slug
     });
     createNodeField({
       node: currNode,
-      name: "prevTitle",
+      name: 'prevTitle',
       value: prevNode.frontmatter.title
     });
     createNodeField({
       node: currNode,
-      name: "prevSlug",
+      name: 'prevSlug',
       value: prevNode.fields.slug
     });
   }
 }
 
-exports.onCreateNode = ({ node, boundActionCreators, getNode }) => {
-  const { createNodeField } = boundActionCreators;
+exports.onCreateNode = ({ node, actions, getNode }) => {
+  const { createNodeField } = actions;
   let slug;
-  if (node.internal.type === "MarkdownRemark") {
+  if (node.internal.type === 'MarkdownRemark') {
     const fileNode = getNode(node.parent);
     const parsedFilePath = path.parse(fileNode.relativePath);
     if (
-      Object.prototype.hasOwnProperty.call(node, "frontmatter") &&
-      Object.prototype.hasOwnProperty.call(node.frontmatter, "title")
+      Object.prototype.hasOwnProperty.call(node, 'frontmatter') &&
+      Object.prototype.hasOwnProperty.call(node.frontmatter, 'title')
     ) {
       slug = `/${_.kebabCase(node.frontmatter.title)}`;
-    } else if (parsedFilePath.name !== "index" && parsedFilePath.dir !== "") {
+    } else if (parsedFilePath.name !== 'index' && parsedFilePath.dir !== '') {
       slug = `/${parsedFilePath.dir}/${parsedFilePath.name}/`;
-    } else if (parsedFilePath.dir === "") {
+    } else if (parsedFilePath.dir === '') {
       slug = `/${parsedFilePath.name}/`;
     } else {
       slug = `/${parsedFilePath.dir}/`;
     }
-    if (
-      Object.prototype.hasOwnProperty.call(node, "frontmatter") &&
-      Object.prototype.hasOwnProperty.call(node.frontmatter, "slug")
-    ) {
-      slug = `/${_.kebabCase(node.frontmatter.slug)}`;
+
+    if (Object.prototype.hasOwnProperty.call(node, 'frontmatter')) {
+      if (Object.prototype.hasOwnProperty.call(node.frontmatter, 'slug'))
+        slug = `/${_.kebabCase(node.frontmatter.slug)}`;
+      if (Object.prototype.hasOwnProperty.call(node.frontmatter, 'date')) {
+        const date = moment(node.frontmatter.date, siteConfig.dateFromFormat);
+        if (!date.isValid) console.warn(`WARNING: Invalid date.`, node.frontmatter);
+
+        createNodeField({
+          node,
+          name: 'date',
+          value: date.toISOString()
+        });
+      }
     }
-    createNodeField({ node, name: "slug", value: slug });
+    createNodeField({ node, name: 'slug', value: slug });
     postNodes.push(node);
   }
 };
 
-exports.setFieldsOnGraphQLNodeType = ({ type, boundActionCreators }) => {
+exports.setFieldsOnGraphQLNodeType = ({ type, actions }) => {
   const { name } = type;
-  const { createNodeField } = boundActionCreators;
-  if (name === "MarkdownRemark") {
+  const { createNodeField } = actions;
+  if (name === 'MarkdownRemark') {
     addSiblingNodes(createNodeField);
   }
 };
 
-exports.createPages = ({ graphql, boundActionCreators }) => {
-  const { createPage } = boundActionCreators;
+exports.createPages = ({ graphql, actions }) => {
+  const { createPage } = actions;
 
   return new Promise((resolve, reject) => {
-    const postPage = path.resolve("src/templates/post.jsx");
-    const tagPage = path.resolve("src/templates/tag.jsx");
-    const categoryPage = path.resolve("src/templates/category.jsx");
+    const postPage = path.resolve('src/templates/post.jsx');
+    const tagPage = path.resolve('src/templates/tag.jsx');
+    const categoryPage = path.resolve('src/templates/category.jsx');
     resolve(
       graphql(
         `
@@ -154,10 +170,4 @@ exports.createPages = ({ graphql, boundActionCreators }) => {
       })
     );
   });
-};
-
-exports.modifyWebpackConfig = ({ config, stage }) => {
-  if (stage === "build-javascript") {
-    config.plugin("Lodash", webpackLodashPlugin, null);
-  }
 };
