@@ -1,11 +1,11 @@
-import React, { Component } from "react";
-import PostListing from "../../components/PostListing/PostListing";
+import React, { Component } from 'react';
+import PostListing from '../PostListing/PostListing';
 
 export default class Archives extends Component {
   render() {
     const { postEdges } = this.props;
     return (
-      <div>
+      <div className="animated zoomInUp">
         <PostListing postEdges={postEdges} />
       </div>
     );
