@@ -3,10 +3,10 @@ import "./AvatarImage.scss";
 
 export default class AvatarImage extends Component {
   render() {
-    const { alt, src } = this.props;
+    const { alt, animation = 'fadeInLeft', src } = this.props;
     return (
       <div className="avatarImage-wrapper wrapper-flex-column--center">
-        <img className="animated fadeInLeft" src={src} alt={alt} />
+        <img className={`animated ${animation}`} src={src} alt={alt} />
       </div>
     );
   }
