@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import Button from "react-md/lib/Buttons";
-import Link from "gatsby-link";
+import { Link } from "gatsby";
 import config from "../../../data/SiteConfig";
 import "./Footer.scss";
 
