@@ -1,9 +1,11 @@
-import React from "react";
-import ObservationListItem from "../ObservationListItem/ObservationListItem";
+import React from 'react';
+import ObservationListItem from '../ObservationListItem/ObservationListItem';
 
 const ObservationList = ({ observations }) => (
   <section>
-    {observations.map(observation => <ObservationListItem key={observation.id} observation={observation} />)}
+    {observations.map(observation => (
+      <ObservationListItem key={observation.id} observation={observation} />
+    ))}
   </section>
 );
 
