@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import FontIcon from "react-md/lib/FontIcons";
-import Link from "gatsby-link";
+import { Link } from "gatsby";
 import "./PostSuggestions.scss";
 
 export default class PostSuggestions extends Component {
