@@ -1,34 +1,31 @@
-import React from 'react';
+import React, { Component } from 'react';
 import Helmet from 'react-helmet';
-
 import { graphql } from 'gatsby';
-import SEO from '../components/SEO/SEO';
-import Home from '../components/Home/Home';
 
+import FourOhFour from '../components/FourOhFour/FourOhFour';
 import config from '../../data/SiteConfig';
 
-class Index extends React.Component {
+class FourOhFourPage extends Component {
   render() {
     const postEdges = this.props.data.allMarkdownRemark.edges;
 
     return (
-      <div className="index-container">
+      <div className="about-container">
         <Helmet>
-          <title>{config.siteTitle}</title>
-          <link rel="canonical" href={`${config.siteUrl}`} />
+          <title>{`About Me | ${config.siteTitle}`}</title>
+          <link rel="canonical" href={`${config.siteUrl}/about-me/`} />
         </Helmet>
-        <SEO postEdges={postEdges} />
-        <Home postEdges={postEdges} />
+        <FourOhFour postEdges={postEdges} />
       </div>
     );
   }
 }
 
-export default Index;
+export default FourOhFourPage;
 
 /* eslint no-undef: "off" */
 export const pageQuery = graphql`
-  query IndexQuery {
+  query FourOhFourQuery {
     allMarkdownRemark(limit: 10, sort: { fields: [frontmatter___date], order: DESC }) {
       edges {
         node {
