@@ -1,22 +1,19 @@
-import React, { Component } from "react";
-import Helmet from "react-helmet";
-import PageTransition from "gatsby-plugin-page-transitions";
+import React, { Component } from 'react';
+import Helmet from 'react-helmet';
 
-import Resume from "../../components/Resume/Resume";
-import config from "../../../data/SiteConfig";
+import Resume from '../../components/Resume/Resume';
+import config from '../../../data/SiteConfig';
 
 class ResumePage extends Component {
   render() {
     return (
-      <PageTransition>
-        <div>
-          <Helmet>
-            <title>{`Resume | ${config.siteTitle}`}</title>
-            <link rel="canonical" href={`${config.siteUrl}/professional/resume/`} />
-          </Helmet>
-          <Resume />
-        </div>
-      </PageTransition>
+      <div>
+        <Helmet>
+          <title>{`Resume | ${config.siteTitle}`}</title>
+          <link rel="canonical" href={`${config.siteUrl}/professional/resume/`} />
+        </Helmet>
+        <Resume />
+      </div>
     );
   }
 }
