@@ -1,22 +1,19 @@
-import React, { Component } from "react";
-import Helmet from "react-helmet";
-import PageTransition from "gatsby-plugin-page-transitions";
+import React, { Component } from 'react';
+import Helmet from 'react-helmet';
 
-import Services from "../../../src/components/Services/Services";
-import config from "../../../data/SiteConfig";
+import Services from '../../components/Services/Services';
+import config from '../../../data/SiteConfig';
 
 class ServicesPage extends Component {
   render() {
     return (
-      <PageTransition>
-        <div>
-          <Helmet>
-            <title>{`Services | ${config.siteTitle}`}</title>
-            <link rel="canonical" href={`${config.siteUrl}/professional/services/`} />
-          </Helmet>
-          <Services />
-        </div>
-      </PageTransition>
+      <div>
+        <Helmet>
+          <title>{`Services | ${config.siteTitle}`}</title>
+          <link rel="canonical" href={`${config.siteUrl}/professional/services/`} />
+        </Helmet>
+        <Services />
+      </div>
     );
   }
 }
