@@ -1,22 +1,19 @@
-import React, { Component } from "react";
-import Helmet from "react-helmet";
-import PageTransition from "gatsby-plugin-page-transitions";
+import React, { Component } from 'react';
+import Helmet from 'react-helmet';
 
-import DevelopmentResources from "../../components/DevelopmentResources/DevelopmentResources";
-import config from "../../../data/SiteConfig";
+import DevelopmentResources from '../../components/DevelopmentResources/DevelopmentResources';
+import config from '../../../data/SiteConfig';
 
 class DevelopmentResourcesPage extends Component {
   render() {
     return (
-      <PageTransition>
-        <div className="">
-          <Helmet>
-            <title>{`Development Resources | ${config.siteTitle}`}</title>
-            <link rel="canonical" href={`${config.siteUrl}/resources/development/`} />
-          </Helmet>
-          <DevelopmentResources />
-        </div>
-      </PageTransition>
+      <div className="">
+        <Helmet>
+          <title>{`Development Resources | ${config.siteTitle}`}</title>
+          <link rel="canonical" href={`${config.siteUrl}/resources/development/`} />
+        </Helmet>
+        <DevelopmentResources />
+      </div>
     );
   }
 }
