@@ -1,7 +1,8 @@
-import React from "react";
-import Helmet from "react-helmet";
-import PostListing from "../components/PostListing/PostListing";
-import config from "../../data/SiteConfig";
+import React from 'react';
+import Helmet from 'react-helmet';
+import { graphql } from 'gatsby';
+import PostListing from '../components/PostListing/PostListing';
+import config from '../../data/SiteConfig';
 
 export default class CategoryTemplate extends React.Component {
   render() {
@@ -10,13 +11,8 @@ export default class CategoryTemplate extends React.Component {
     return (
       <div className="category-container">
         <Helmet>
-          <title>
-            {`Posts in category "${category}" | ${config.siteTitle}`}
-          </title>
-          <link
-            rel="canonical"
-            href={`${config.siteUrl}/categories/${category}`}
-          />
+          <title>{`Posts in category "${category}" | ${config.siteTitle}`}</title>
+          <link rel="canonical" href={`${config.siteUrl}/categories/${category}`} />
         </Helmet>
         <PostListing postEdges={postEdges} />
       </div>
