@@ -1,7 +1,9 @@
-import React from "react";
-import Helmet from "react-helmet";
-import PostListing from "../components/PostListing/PostListing";
-import config from "../../data/SiteConfig";
+import React from 'react';
+import Helmet from 'react-helmet';
+import { graphql } from 'gatsby';
+
+import PostListing from '../components/PostListing/PostListing';
+import config from '../../data/SiteConfig';
 
 export default class TagTemplate extends React.Component {
   render() {
